requirements.txt
packaging==20.4
python-engineio==3.13.2
python-socketio==4.6.0
SoundFile==0.10.3.post1
Flask==1.1.2
Flask-Cors==3.0.9
Flask-SocketIO==4.3.1
grpcio-tools==1.32.0
grpcio==1.32.0
inotify
google-cloud-core==1.4.3
google-cloud-storage==1.31.2
pydub==0.24.1
